+++ bindsnet/__init__.py
@@ -1,18 +1,18 @@
 from pathlib import Path
 
-__all__ = [
-    "utils",
-    "network",
-    "models",
-    "analysis",
-    "preprocessing",
-    "datasets",
-    "encoding",
-    "pipeline",
-    "learning",
-    "evaluation",
-    "environment",
-    "conversion",
-]
+from . import (
+    utils,
+    network,
+    models,
+    analysis,
+    preprocessing,
+    datasets,
+    encoding,
+    pipeline,
+    learning,
+    evaluation,
+    environment,
+    conversion,
+)
 
 ROOT_DIR = Path(__file__).parents[0].parents[0]
