+++ bindsnet/analysis/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["plotting", "visualization", "pipeline_analysis"]
+from . import plotting, visualization, pipeline_analysis
