+++ bindsnet/encoding/loaders.py
@@ -1,9 +1,8 @@
 from typing import Optional, Union, Iterable, Iterator
 
 import torch
-import numpy as np
 
-from .encodings import single, repeat, bernoulli, poisson, rank_order
+from .encodings import bernoulli, poisson, rank_order
 
 
 def bernoulli_loader(
@@ -29,9 +28,8 @@ def bernoulli_loader(
     max_prob = kwargs.get("dt", 1.0)
 
     for i in range(len(data)):
-        yield bernoulli(
-            datum=data[i], time=time, dt=dt, max_prob=max_prob
-        )  # Encode datum as Bernoulli spike trains.
+        # Encode datum as Bernoulli spike trains.
+        yield bernoulli(datum=data[i], time=time, dt=dt, max_prob=max_prob)
 
 
 def poisson_loader(
@@ -50,9 +48,8 @@ def poisson_loader(
     :return: Tensors of shape ``[time, n_1, ..., n_k]`` of Poisson-distributed spikes.
     """
     for i in range(len(data)):
-        yield poisson(
-            datum=data[i], time=time, dt=dt
-        )  # Encode datum as Poisson spike trains.
+        # Encode datum as Poisson spike trains.
+        yield poisson(datum=data[i], time=time, dt=dt)
 
 
 def rank_order_loader(
@@ -71,6 +68,5 @@ def rank_order_loader(
     :return: Tensors of shape ``[time, n_1, ..., n_k]`` of rank order-encoded spikes.
     """
     for i in range(len(data)):
-        yield rank_order(
-            datum=data[i], time=time, dt=dt
-        )  # Encode datum as rank order-encoded spike trains.
+        # Encode datum as rank order-encoded spike trains.
+        yield rank_order(datum=data[i], time=time, dt=dt)
