+++ bindsnet/network/__init__.py
@@ -1,3 +1,2 @@
 from .network import Network, load
-
-__all__ = ["load", "Network", "nodes", "monitors", "topology"]
+from . import nodes, topology, monitors
