+++ bindsnet/pipeline/__init__.py
@@ -1,11 +1,4 @@
 from .environment_pipeline import EnvironmentPipeline
 from .base_pipeline import BasePipeline
 from .dataloader_pipeline import DataLoaderPipeline, TorchVisionDatasetPipeline
-
-__all__ = [
-    "BasePipeline",
-    "EnvironmentPipeline",
-    "DataLoaderPipeline",
-    "TorchVisionDatasetPipeline",
-    "action",
-]
+from . import action
